Move Flatten and FlatMap to own module
